Modify storage interface, add new storage impl
- Change capacity and item size return type to size_t
- Fix warning: Add variable to lock guard instead of pure constructor calling
- Modify fill function in Pool

+++ FactoryMethod/Projects/Logistics/src/storage/Storage.h
@@ -1,3 +1,7 @@
+//
+// Created by bellkross on 24.02.2019.
+//
+
 #ifndef STORAGE_H_
 #define STORAGE_H_
 
@@ -13,8 +17,8 @@ public:
 	virtual T* getResource() = 0;
 	virtual void returnResource(T* resource) = 0;
 	virtual bool isEmpty() const = 0;
-	virtual int getCapacity() const = 0;
-	virtual int getItemsCount() const = 0;
+	virtual size_t getCapacity() const = 0;
+	virtual size_t getItemsCount() const = 0;
 	virtual ~Storage() = default;
 };
